+++ src/_components/AudioElement/AudioElement.tsx
@@ -1,0 +1,5 @@
+import styles from "./AudioElement.module.css";
+
+export default function AudioElement() {
+  return <div className={styles["player"]}></div>;
+}